Struct gsdk::Value

pub struct Value<T = ()> {
    pub value: ValueDef<T>,
    pub context: T,
}

Value holds a representation of some value that has been decoded, as well as some arbitrary context.

Not all SCALE encoded types have an similar-named value; for instance, the values corresponding to sequence, array and composite types can all be represented with [Composite]. Only enough information is preserved here to to be able to encode and decode SCALE bytes with a known type to and from Values losslessly.

Fields
value: ValueDef<T>

The shape and associated data for this Value

context: T

Some additional arbitrary context that can be associated with a value.

Implementations

impl Value

pub fn named_composite<S, Vals>(vals: Vals) -> Value
where S: Into<String>, Vals: IntoIterator<Item = (S, Value)>,

Construct a named composite type from any type which produces a tuple of keys and values when iterated over.

pub fn unnamed_composite<Vals>(vals: Vals) -> Value
where Vals: IntoIterator<Item = Value>,

Construct an unnamed composite type from any type which produces values when iterated over.

pub fn variant<S>(name: S, values: Composite<()>) -> Value
where S: Into<String>,

Create a new variant value without additional context.

pub fn named_variant<S, F, Vals>(name: S, fields: Vals) -> Value
where S: Into<String>, F: Into<String>, Vals: IntoIterator<Item = (F, Value)>,

Create a new variant value with named fields and without additional context.

pub fn unnamed_variant<S, Vals>(name: S, fields: Vals) -> Value
where S: Into<String>, Vals: IntoIterator<Item = Value>,

Create a new variant value with tuple-like fields and without additional context.

pub fn bit_sequence(bits: Bits) -> Value

Create a new bit sequence value without additional context.

pub fn primitive(primitive: Primitive) -> Value

Create a new primitive value without additional context.

pub fn string<S>(val: S) -> Value
where S: Into<String>,

Create a new string value without additional context.

pub fn bool(val: bool) -> Value

Create a new boolean value without additional context.

pub fn char(val: char) -> Value

Create a new char without additional context.

pub fn u128(val: u128) -> Value

Create a new unsigned integer without additional context.

pub fn i128(val: i128) -> Value

Create a new signed integer without additional context.

pub fn from_bytes(bytes: impl AsRef<[u8]>) -> Value

Create a new Value from a set of bytes; useful for converting things like AccountIds.

impl Value

pub fn without_context(value: ValueDef<()>) -> Value

Create a new value with no associated context.

impl<T> Value<T>

pub fn with_context(value: ValueDef<T>, context: T) -> Value<T>

Create a new value with some associated context.

pub fn remove_context(self) -> Value

Remove the context.

pub fn map_context<F, U>(self, f: F) -> Value<U>
where F: Clone + FnMut(T) -> U,

Map the context to some different type.

pub fn as_bool(&self) -> Option<bool>

If the value is a boolean value, return it.

pub fn as_char(&self) -> Option<char>

If the value is a char, return it.

pub fn as_u128(&self) -> Option<u128>

If the value is a u128, return it.

pub fn as_i128(&self) -> Option<i128>

If the value is an i128, return it.

pub fn as_str(&self) -> Option<&str>

If the value is a string, return it.