Crate gsdk

Gear api

Re-exports

pub use crate::config::GearConfig;

pub use crate::metadata::Event;

pub use crate::result::Error;

pub use crate::result::Result;

pub use crate::signer::PairSigner;

pub use crate::subscription::Blocks;

pub use crate::subscription::Events;

Modules

backtrace

Backtrace support for gsdk

config

Gear config

events

Events api

ext

gp

generated code preludes.

metadata

result

GSdk Results

signer

Gear api with signer

subscription

Subscription implementation.

Structs

Api

Gear api wrapper.

ApiBuilder

gsdk api builder

GasInfo

Pre-calculated gas consumption estimate for a message.

Value

Value holds a representation of some value that has been decoded, as well as some arbitrary context.

Enums

Program

Gear Program

Type Aliases

BlockNumber

Block number type

GearGasNode

Gear gas node.

GearGasNodeId

Gear gas node id.

GearPages

Gear pages.

TxInBlock

Transaction in block.

TxInBlockResult

Transaction in block with result wrapper.

TxStatus

Transaction status.