Struct gsdk::Api

pub struct Api {
    pub retry: u16,
    /* private fields */
}

Gear api wrapper.

Fields
retry: u16

How many times we’ll retry when rpc requests failed.

Implementations

impl Api

pub fn gas_limit(&self) -> Result<u64>

pallet gas constants

Get gas limit.

pub fn expected_block_time(&self) -> Result<u64>

pallet babe constant

Get expected block time.

impl Api

pub async fn capture_dispatch_info( &self, tx: &TxInBlock<GearConfig, OnlineClient<GearConfig>> ) -> Result<TxEvents<GearConfig>>

Capture the dispatch info of any extrinsic and display the weight spent

pub fn capture_weight_info(details: &EventDetails) -> Result<()>

Parse transaction fee from InBlockEvents

impl Api

pub async fn calculate_create_gas( &self, origin: H256, code_id: CodeId, payload: Vec<u8>, value: u128, allow_other_panics: bool, at: Option<H256> ) -> Result<GasInfo>

gear_calculateInitCreateGas

pub async fn calculate_upload_gas( &self, origin: H256, code: Vec<u8>, payload: Vec<u8>, value: u128, allow_other_panics: bool, at: Option<H256> ) -> Result<GasInfo>

gear_calculateInitUploadGas

pub async fn calculate_handle_gas( &self, origin: H256, destination: ProgramId, payload: Vec<u8>, value: u128, allow_other_panics: bool, at: Option<H256> ) -> Result<GasInfo>

gear_calculateHandleGas

pub async fn calculate_reply_gas( &self, origin: H256, message_id: MessageId, exit_code: i32, payload: Vec<u8>, value: u128, allow_other_panics: bool, at: Option<H256> ) -> Result<GasInfo>

gear_calculateReplyGas

impl Api

pub async fn fetch_storage<'a, Address, Value>( &self, address: &'a Address ) -> Result<Value>where Address: StorageAddress<IsFetchable = Yes, IsDefaultable = Yes, Target = DecodedValueThunk> + 'a, Value: Decode,

Shortcut for fetching storage.

pub async fn program_pages(&self, pid: H256) -> Result<GearPages>

Get program pages from program id.

impl Api

pub async fn info( &self, address: &str ) -> Result<AccountInfo<u32, AccountData<u128>>>

Get account info by address

pub async fn number(&self) -> Result<u32>

Get block number.

pub async fn get_balance(&self, address: &str) -> Result<u128>

Get balance by account address

pub async fn get_events_at( &self, block_hash: Option<H256> ) -> Result<Vec<RuntimeEvent>>

impl Api

pub async fn block_timestamp(&self, block_hash: Option<H256>) -> Result<u64>

Return a timestamp of the block.

impl Api

pub async fn validators(&self) -> Result<Vec<AccountId32>>

Get all validators from pallet_session.

impl Api

pub async fn execute_inherent(&self) -> Result<bool>

Check whether the message queue processing is stopped or not.

pub async fn gear_block_number(&self, block_hash: Option<H256>) -> Result<u32>

Get gear block number.

impl Api

pub async fn code_storage(&self, code_hash: [u8; 32]) -> Result<InstrumentedCode>

Get InstrumentedCode by code_hash

pub async fn gprog(&self, pid: H256) -> Result<ActiveProgram>

Get active program from program id.

pub async fn gpages( &self, pid: H256, program: &ActiveProgram ) -> Result<GearPages>

Get pages of active program.

impl Api

pub async fn get_from_account_mailbox( &self, account_id: AccountId32, message_id: impl AsRef<[u8]> ) -> Result<Option<(StoredMessage, Interval<u32>)>>

Get a message identified by message_id from the account_id’s mailbox.

pub async fn mailbox( &self, address: AccountId32, count: u32 ) -> Result<Vec<(StoredMessage, Interval<u32>)>>

Get mailbox from address

impl Api

pub fn cmp_gas_limit(&self, gas: u64) -> Result<u64>

compare gas limit

pub fn decode_error(&self, dispatch_error: DispatchError) -> Error

Decode DispatchError to subxt::error::Error.

impl Api

pub async fn new(url: Option<&str>) -> Result<Self>

Create new API client.

pub async fn new_with_timeout( url: Option<&str>, timeout: Option<u64> ) -> Result<Self>

Create new API client with timeout.

pub fn with_retry(self, retry: u16) -> Self

Setup retry times and return the API instance.

pub async fn blocks(&self) -> Result<Blocks>

Subscribe all blocks

pub async fn finalized_blocks(&self) -> Result<Blocks>

Subscribe finalized blocks

pub fn signer(self, suri: &str, passwd: Option<&str>) -> Result<Signer>

New signer from api